+++ pytorch_lightning/utilities/data.py
@@ -95,18 +95,18 @@ def has_len_all_ranks(
 
         if total_length == 0:
             raise MisconfigurationException(
-                " Total length of `Dataloader` across ranks is zero. Please make sure that it returns at least 1 batch"
+                "Total length of `Dataloader` across ranks is zero. Please make sure that it returns at least 1 batch."
             )
         if total_length > 0 and local_length == 0:
             if model.allow_zero_length_dataloader_with_multiple_devices:
                 rank_zero_warn(
-                    "Total length of `Dataloader` across ranks is zero, but local rank has zero length. "
-                    "Please be cautious of uneven batch length. "
+                    "Total length of `Dataloader` across ranks is zero, but local rank has zero length."
+                    " Please be cautious of uneven batch length."
                 )
                 has_len = False
             else:
                 raise MisconfigurationException(
-                    "`Dataloader` within local rank has zero length. Please make sure that it returns at least 1 batch"
+                    "`Dataloader` within local rank has zero length. Please make sure that it returns at least 1 batch."
                 )
         else:
             has_len = True